feat: add loans management plugin with CRUD operations and UI components

+++ plugins/books-plugin/src/main/java/br/edu/ifba/inf008/plugins/books/domain/exceptions/BookWithLockedCopiesException.java
@@ -5,7 +5,7 @@ package br.edu.ifba.inf008.plugins.books.domain.exceptions;
  */
 public class BookWithLockedCopiesException extends Exception {
 
-    public BookWithLockedCopiesException(Long bookId, Integer currentCopies) {
+    public BookWithLockedCopiesException(Long bookId, Long currentCopies) {
         super("The book with ID " + bookId + " has locked copies: " + currentCopies + ". " +
               "Please unlock the copies before proceeding.");
     }
