+++ plugins/books-plugin/src/main/java/br/edu/ifba/inf008/plugins/books/infrastructure/providers/BookAvailabilityProvider.java
@@ -1,5 +1,5 @@
 package br.edu.ifba.inf008.plugins.books.infrastructure.providers;
 
 public interface BookAvailabilityProvider {
-    Integer provideAvailableCopies(Long bookId);
+    Long provideAvailableCopies(Long bookId);
 }
